move linkflow into dir
