+++ packages/wallet-sdk/src/provider/WalletSDKUI.ts
@@ -1,4 +1,4 @@
-import { LinkFlow } from "../components/LinkFlow";
+import { LinkFlow } from "../components/LinkFlow/LinkFlow";
 import { Snackbar, SnackbarInstanceProps } from "../components/Snackbar/Snackbar";
 import { injectCssReset } from "../lib/cssReset";
 import {
